backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
chromadb==0.4.18
sentence-transformers==2.2.2
pydantic==2.5.2
python-multipart==0.0.6
bcrypt==4.0.1
python-jose==3.3.0
passlib==1.7.4
langchain==0.0.352
langchain-google-genai==0.0.5
faiss-cpu==1.7.4
